product page + user login

+++ src/index.js
@@ -9,7 +9,7 @@ import {initializeApp} from 'https://www.gstatic.com/firebasejs/10.5.0/firebase-
 
 import { collection, getFirestore, onSnapshot,
   addDoc,deleteDoc,doc,query,where, orderBy,
-  serverTimestamp,getDoc,updateDoc } from 'https://www.gstatic.com/firebasejs/10.5.0/firebase-firestore.js';
+  serverTimestamp,getDoc,updateDoc} from 'https://www.gstatic.com/firebasejs/10.5.0/firebase-firestore.js';
 
 import { getAuth,createUserWithEmailAndPassword,
   signOut,signInWithEmailAndPassword,onAuthStateChanged } from 'https://www.gstatic.com/firebasejs/10.5.0/firebase-auth.js';
@@ -26,8 +26,9 @@ const firebaseConfig = {
 const app = initializeApp(firebaseConfig); // inicjalizacja połączenia do backendu Firebase
 
 const db = getFirestore();
-
+const auth = getAuth();
 const productsColRef = collection(db,'products');
+const userColRef = collection(db,'users');
 
-export {app, db, productsColRef, onSnapshot, addDoc, deleteDoc, doc, query, where, orderBy, serverTimestamp, getDoc, updateDoc};
+export {app, db, productsColRef,userColRef, onSnapshot, addDoc, deleteDoc, doc, query, where, orderBy, serverTimestamp, getDoc, updateDoc, getAuth, createUserWithEmailAndPassword, signOut, signInWithEmailAndPassword, onAuthStateChanged,auth};
